attachment changes

+++ resources/views/projects/attachmentAjax.blade.php
@@ -5,9 +5,9 @@
  				<table class="table table-striped table-bordered first">
  					<thead>
  						<tr>
- 							<th>Heading</th>
- 							<th>Heading</th>
- 							<th>Documents</th>
+ 							<th>SR No</th>
+ 							<th>Attachment Name</th>
+ 							<th>Attachment File</th>
  							<th>Action</th>
 
  						</tr>
@@ -17,8 +17,12 @@
  						<tr>
  							<td>{{$value['id']}}</td>
  							<td>{{$value['heading']}}</td>
- 							<td>{{$value['documents']}}</td>
- 							<td><a href="#" rel="noopener noreferrer" data-attachment ="{{json_encode($value)}}" class="editAttachment" title="Edit">
+ 							<td>
+ 								<a href="{{ asset('images/attachment/' . $value['project_id'] . '/' . $value['documents']) }}" traget="_blank">
+ 									{{ $value['documents'] }}
+ 								</a>
+ 							</td>
+ 							<td><a href="#" rel="noopener noreferrer" data-attachment="{{json_encode($value)}}" class="editAttachment" title="Edit">
  									<i class="fas fa-edit text-primary" style="font-size: 1rem"></i>
  								</a>
  								<a href="#" class="ml-2" onclick="removeLab('{{ $value['id'] }}')" title="Delete">
@@ -38,37 +42,57 @@
  </div>
  @push('js')
  <script>
-$(document).on('click', '.editAttachment', function() {
-	var attachment = $(this).data('attachment');
+ 	$(document).on('click', '.editAttachment', function() {
+ 		var attachment = $(this).data('attachment');
 
-	$("#id").val(attachment.id);
-	$("#heading").val(attachment.heading);
-$(".documentsValue").show();
-$(".documentsValue").text(attachment.documents);
+ 		$("#id").val(attachment.id);
+ 		$("#heading").val(attachment.heading);
+ 		$(".documentsValue").show();
+ 		$(".documentsValue").text(attachment.documents);
 
-	$("#attachmentModel").modal('show');
+ 		$("#attachmentModel").modal('show');
 
-	console.log(attachment.id);
-	
-});
+ 		console.log(attachment.id);
+
+ 	});
 
  	function removeLab(id) {
- 		$.ajax({
- 			type: 'DELETE',
- 			url: "{{ url('attachment/remove/') }}" + "/" + id,
- 			headers: {
- 				'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
+ 		swal({
+ 				title: "Are you sure?",
+ 				text: "You will not be able to recover this imaginary file!",
+ 				type: "warning",
+ 				showCancelButton: true,
+ 				confirmButtonColor: "#DD6B55",
+ 				confirmButtonText: "Yes, delete it!",
+ 				cancelButtonText: "No, cancel plx!",
+ 				closeOnConfirm: false,
+ 				closeOnCancel: false
  			},
- 			success: function(response) {
- 				if (response.status) {
- 					$(".loadView").html(response.html)
- 					$("#attachmentModel").modal('hide');
+ 			function(isConfirm) {
+ 				if (isConfirm) {
+ 					$.ajax({
+ 						type: 'DELETE',
+ 						url: "{{ url('attachment/remove/') }}" + "/" + id,
+ 						headers: {
+ 							'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
+ 						},
+ 						success: function(response) {
+ 							if (response.status) {
+ 								$(".loadView").html(response.html)
+ 								$("#attachmentModel").modal('hide');
+
+ 							}
+ 						},
+ 						error: function(xhr, status, error) {
+ 							console.error(xhr.responseText);
+ 						}
+ 					});
+ 					swal("Deleted!", "Your imaginary file has been deleted.", "success");
+ 				} else {
+ 					swal("Cancelled", "Your imaginary file is safe :)", "error");
  				}
- 			},
- 			error: function(xhr, status, error) {
- 				console.error(xhr.responseText);
- 			}
- 		});
+ 			});
+
  	}
  </script>
  @endpush
